refactor(test/storybook): introduce UserModule.mock.ts
Aligns better with emerging naming convention for mocks where the data
being mocked is in mock files named after the module etc that manages
the live data.
