+++ react-front-end/__mocks__/UserModule.mock.ts
@@ -17,6 +17,42 @@
  */
 import * as OEQ from "@openequella/rest-api-client";
 
+/**
+ * A list of users to test with, deliberately out of order. Names are randomly generated.
+ */
+export const users: OEQ.UserQuery.UserDetails[] = [
+  {
+    id: "680f5eb7-22e2-4ab6-bcea-25205165e36e",
+    username: "user200",
+    firstName: "Fabienne",
+    lastName: "Hobson",
+  },
+  {
+    id: "cda09b86-3662-46bd-b60e-4bce89efba7a",
+    username: "user100",
+    firstName: "Racheal",
+    lastName: "Carlyle",
+  },
+  {
+    id: "97254515-6e32-48e9-ba65-5b5c6aa182a6",
+    username: "user400",
+    firstName: "Ronny",
+    lastName: "Southgate",
+  },
+  {
+    id: "8db50158-757d-44f3-8ccf-7b2e0a3a6405",
+    username: "user300",
+    firstName: "Yasmin",
+    lastName: "Day",
+  },
+  {
+    id: "3f25f543-7231-46f4-8f3b-aaccc8fcf52a",
+    username: "admin999",
+    firstName: "Wat",
+    lastName: "Swindlehurst",
+  },
+];
+
 export const getCurrentUserMock: OEQ.LegacyContent.CurrentUserDetails = {
   id: "test",
   username: "test",
@@ -29,3 +65,13 @@ export const getCurrentUserMock: OEQ.LegacyContent.CurrentUserDetails = {
   menuGroups: [],
   canDownloadSearchResult: true,
 };
+
+/**
+ * Helper function to inject into component for user retrieval by an array of ids.
+ *
+ * @param ids A list of user IDs to lookup, should be one of those in `users`
+ */
+export const resolveUsersProvider = async (
+  ids: ReadonlyArray<string>
+): Promise<OEQ.UserQuery.UserDetails[]> =>
+  Promise.resolve(users.filter(({ id }) => ids.includes(id)));
